+++ ChangeLog.md
@@ -1,3 +1,8 @@
+Version 2.1.8 -----------------------------------------------------
+  - Added MD5 Brutforce and Dictionary attack
+  - Fixed some bugs
+  - Now you can drag the file to the console instead of writing the full path to a file in the Offline Decrypt Menu
+
 Version 2.1.7 -----------------------------------------------------
   - Added Number selection for interface selection menu